requirements.txt
# Video generation
litellm>=1.50.0
openai>=1.0.0
fal-client>=0.5.0
httpx>=0.27.0

# Backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12

# Env
python-dotenv>=1.0.0

# Video analysis
opencv-python-headless>=4.8.0
Pillow>=10.0.0
requests>=2.32.0
yt-dlp>=2024.1.1
